requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
Django==3.2.16
django-cart==1.0.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
fontawesomefree==6.2.1
greenlet==2.0.1
idna==3.4
mercadopago==2.2.0
pycparser==2.21
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
typing_extensions==4.4.0
unicode==2.9
urllib3==1.26.13
zope.event==4.5.0
zope.interface==5.5.2
